requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
transformers==4.35.2
torch==2.1.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
aiohttp==3.9.1
tweepy==4.14.0
pymongo==4.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1 